+++ prog01/produto.cpp
@@ -5,7 +5,7 @@ Produto::Produto() {}
 
 Produto::~Produto(){}
 
-Produto::Produto(std::string _codigo, std::string _descricao, short _preco):
+Produto::Produto(std::string _codigo, std::string _descricao, double _preco):
 	m_cod_barras(_codigo), m_descricao(_descricao), m_preco(_preco) {}
 
 std::string
