+++ prog01/produto.h
@@ -7,7 +7,7 @@ class Produto
 {
 public:
 	Produto();
-	Produto(std::string _codigo, std::string _descricao, short _preco);
+	Produto(std::string _codigo, std::string _descricao, double _preco);
 	virtual ~Produto();
 protected:
 	std::string m_cod_barras;
